cloudrun/requirements.txt
Flask==2.0.1
google-cloud-tasks==2.8.1
gunicorn==20.1.0
pendulum==2.1.2
werkzeug==2.1.2
